sistema de categorias adicionado

+++ app/Http/Controllers/DeleteProductController.php
@@ -9,11 +9,16 @@ use Illuminate\Support\Facades\Storage;
 class DeleteProductController extends Controller
 {
     public function index($id, Request $request){
+      $ranking = $request->session()->get('ranking');
+      if ($ranking == 2 || $ranking == null){
+          return redirect('/');
+      }else{
         DB::table('images')->where('product_id', '=', $id)->delete();
         DB::table('products')->where('id', '=', $id)->delete();
         
         //Storage::delete('images.'.$product);
 
       return redirect('/viewProduct');
+      }
     }
 }
